# docs/API/Admin/Admin_swagger.yaml
basePath: /api/v1
definitions:
...
  model.BatchCreateSubAdminFailure:
    properties:
      email:
        type: string
      reason:
        type: string
    type: object
  model.BatchCreateSubAdminsRequest:
    properties:
      emails:
        items:
          type: string
        type: array
      permissionNames:
        items:
          type: string
        type: array
    type: object
  model.BatchCreateSubAdminsResponse:
    properties:
      created:
        items:
          $ref: '#/definitions/model.SubAdminCreateResult'
        type: array
      failed:
        items:
          $ref: '#/definitions/model.BatchCreateSubAdminFailure'
        type: array
    type: object
...
  model.CreateSubAdminRequest:
    properties:
      adminEmail:
        type: string
      adminName:
        type: string
      permissionNames:
        items:
          type: string
        type: array
    type: object
...
  model.HandoverSuperAdminRequest:
    properties:
      newSuperAdminID:
        type: integer
    type: object
  model.Response:
    properties:
      code:
        description: 状态码，和http状态码剥离，以分离网络和业务
        example: 200
        type: integer
      msg:
        description: 信息。可以是NULL，也可以是成功时返回的结构体，也可以是失败时的错误信息
    type: object
...
  model.SubAdminCreateResult:
    properties:
      adminEmail:
        type: string
      adminID:
        type: integer
      adminName:
        type: string
      emailError:
        type: string
      emailSent:
        type: boolean
      tempPassword:
        type: string
    type: object
  model.SubAdminInfo:
    properties:
      adminEmail:
        type: string
      adminID:
        type: integer
      adminName:
        type: string
      isActive:
        type: boolean
      permissionNames:
        items:
          type: string
        type: array
    type: object
...
  model.UpdateSubAdminPermissionsRequest:
    properties:
      permissionNames:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Ubik 系统管理后台服务接口文档
  title: 管理后台 API
  version: "1.0"
paths:
...
  /admin/sub-admins:
    get:
      consumes:
      - application/json
      description: 仅超级管理员可获取全部子管理员及其权限
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  items:
                    $ref: '#/definitions/model.SubAdminInfo'
                  type: array
              type: object
      summary: 获取子管理员列表
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: 仅超级管理员可创建子管理员并分配权限，返回临时密码并尝试发送邮件
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建子管理员请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.SubAdminCreateResult'
              type: object
      summary: 创建子管理员
      tags:
      - Admin
  /admin/sub-admins/{admin_id}:
    delete:
      consumes:
      - application/json
      description: 仅超级管理员可删除子管理员账户
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 子管理员ID
        in: path
        name: admin_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除子管理员
      tags:
      - Admin
  /admin/sub-admins/{admin_id}/disable:
    post:
      consumes:
      - application/json
      description: 仅超级管理员可禁用子管理员账户
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 子管理员ID
        in: path
        name: admin_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 禁用成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 禁用子管理员
      tags:
      - Admin
  /admin/sub-admins/{admin_id}/permissions:
    put:
      consumes:
      - application/json
      description: 仅超级管理员可替换子管理员的权限集合
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 子管理员ID
        in: path
        name: admin_id
        required: true
        type: integer
      - description: 权限更新请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubAdminPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 更新子管理员权限
      tags:
      - Admin
  /admin/sub-admins/batch:
    post:
      consumes:
      - application/json
      - text/csv
      description: 支持 JSON({emails:[...]}) 或 text/csv(首列邮箱) 的批量导入，返回每条创建结果与失败原因
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: JSON批量创建请求
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.BatchCreateSubAdminsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建完成
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.BatchCreateSubAdminsResponse'
              type: object
      summary: 批量创建子管理员
      tags:
      - Admin
  /admin/sub-admins/handover-super:
    post:
      consumes:
      - application/json
      description: 将超级管理员权限交接给目标管理员，并禁用当前超级管理员
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 交接请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.HandoverSuperAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 交接成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 超级管理员交接
      tags:
      - Admin
...
swagger: "2.0"